Add .env in .gitignore file and i splitted the Header component with a the Dumb and smart file, called layout.js
