+++ React/tic-tac/.gitignore
@@ -8,16 +8,18 @@ coverage/
 
 # production
 build/
+.env
 
 # misc
 .DS_Store
+.idea/
 .env.local
 .env.development.local
 .env.test.local
 .env.production.local
 
+#developer
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-.idea/
 .csscomb.json
